requirements.txt
pyautogui
opencv-python
pillow
pytest
python-dotenv
mss
numpy
cryptography